Rename isolations file
